+++ temporian/implementation/numpy/operators/BUILD
@@ -34,10 +34,16 @@ py_library(
 )
 
 py_library(
+    name = "base",
+    srcs = ["base.py"],
+    srcs_version = "PY3",
+)
+
+py_library(
     name = "arithmetic",
     srcs = ["arithmetic.py"],
     srcs_version = "PY3",
-    deps = [
+    deps = [ ":base",
         "//temporian/core/operators:arithmetic",
         "//temporian/implementation/numpy:implementation_lib",
         "//temporian/implementation/numpy/data:event",
@@ -49,6 +55,7 @@ py_library(
     srcs = ["glue.py"],
     srcs_version = "PY3",
     deps = [
+        ":base",
         "//temporian/core/operators:glue",
         "//temporian/implementation/numpy:implementation_lib",
         "//temporian/implementation/numpy/data:event",
@@ -60,7 +67,7 @@ py_library(
     name = "lag",
     srcs = ["lag.py"],
     srcs_version = "PY3",
-    deps = [
+    deps = [ ":base",
         "//temporian/core/data:duration",
         "//temporian/core/operators:lag",
         "//temporian/implementation/numpy:implementation_lib",
@@ -73,7 +80,7 @@ py_library(
     name = "select",
     srcs = ["select.py"],
     srcs_version = "PY3",
-    deps = [
+    deps = [ ":base",
         "//temporian/core/operators:select",
         "//temporian/implementation/numpy:implementation_lib",
         "//temporian/implementation/numpy/data:event",
@@ -84,7 +91,7 @@ py_library(
     name = "propagate",
     srcs = ["propagate.py"],
     srcs_version = "PY3",
-    deps = [
+    deps = [ ":base",
         "//temporian/core/data:duration",
         "//temporian/core/operators:propagate",
         "//temporian/implementation/numpy:implementation_lib",
@@ -98,7 +105,7 @@ py_library(
     name = "sample",
     srcs = ["sample.py"],
     srcs_version = "PY3",
-    deps = [
+    deps = [ ":base",
         "//temporian/core/data:duration",
         "//temporian/core/operators:sample",
         "//temporian/implementation/numpy:implementation_lib",
@@ -114,7 +121,7 @@ py_library(
     name = "prefix",
     srcs = ["prefix.py"],
     srcs_version = "PY3",
-    deps = [
+    deps = [ ":base",
         "//temporian/core/data:duration",
         "//temporian/core/operators:prefix",
         "//temporian/implementation/numpy:implementation_lib",
