+++ temporian/implementation/numpy/operators/glue.py
@@ -22,12 +22,13 @@ from temporian.implementation.numpy.data.event import NumpyEvent
 from temporian.implementation.numpy.data.event import NumpyFeature
 from temporian.core.operators.glue import GlueOperator
 from temporian.implementation.numpy import implementation_lib
+from temporian.implementation.numpy.operators.base import OperatorImplementation
 
 
-class GlueNumpyImplementation:
+class GlueNumpyImplementation(OperatorImplementation):
     def __init__(self, operator: GlueOperator):
+        super().__init__(operator)
         assert isinstance(operator, GlueOperator)
-        self._operator = operator
 
     def __call__(
         self,
