+++ temporian/implementation/numpy/operators/select.py
@@ -3,13 +3,14 @@ from typing import Dict
 from temporian.core.operators.select import SelectOperator
 from temporian.implementation.numpy.data.event import NumpyEvent
 from temporian.implementation.numpy import implementation_lib
+from temporian.implementation.numpy.operators.base import OperatorImplementation
 
 
-class SelectNumpyImplementation:
+class SelectNumpyImplementation(OperatorImplementation):
     """Select a subset of features from an event."""
 
     def __init__(self, operator: SelectOperator) -> None:
-        self.operator = operator
+        super().__init__(operator)
 
     def __call__(self, event: NumpyEvent) -> Dict[str, NumpyEvent]:
         feature_names = self.operator.attributes()["feature_names"]
