+++ temporian/implementation/numpy/operators/test/calendar_day_of_month_test.py
@@ -27,7 +27,6 @@ from temporian.implementation.numpy.operators.calendar.day_of_month import (
     CalendarDayOfMonthNumpyImplementation,
 )
 from temporian.core.data import dtype
-from temporian.implementation.numpy.evaluator import run_with_check
 
 
 class CalendarDayOfMonthNumpyImplementationTest(absltest.TestCase):
@@ -63,7 +62,7 @@ class CalendarDayOfMonthNumpyImplementationTest(absltest.TestCase):
 
         operator = CalendarDayOfMonthOperator(input_event)
         impl = CalendarDayOfMonthNumpyImplementation(operator)
-        output = run_with_check(operator, impl, {"sampling": input_event_data})
+        output = impl.call(sampling=input_event_data)
 
         self.assertTrue(output_event_data == output["event"])
         self.assertTrue(
@@ -108,7 +107,7 @@ class CalendarDayOfMonthNumpyImplementationTest(absltest.TestCase):
 
         operator = CalendarDayOfMonthOperator(input_event)
         impl = CalendarDayOfMonthNumpyImplementation(operator)
-        output = run_with_check(operator, impl, {"sampling": input_event_data})
+        output = impl.call(sampling=input_event_data)
 
         self.assertTrue(output_event_data == output["event"])
         self.assertTrue(
