+++ temporian/implementation/numpy/operators/test/calendar_day_of_year_test.py
@@ -25,7 +25,6 @@ from temporian.implementation.numpy.operators.calendar.day_of_year import (
     CalendarDayOfYearNumpyImplementation,
 )
 from temporian.core.data import dtype
-from temporian.implementation.numpy.evaluator import run_with_check
 
 
 class CalendarDayOfYearNumpyImplementationTest(absltest.TestCase):
@@ -65,7 +64,7 @@ class CalendarDayOfYearNumpyImplementationTest(absltest.TestCase):
 
         operator = CalendarDayOfYearOperator(input_event)
         impl = CalendarDayOfYearNumpyImplementation(operator)
-        output = run_with_check(operator, impl, {"sampling": input_event_data})
+        output = impl.call(sampling=input_event_data)
 
         self.assertTrue(output_event_data == output["event"])
         self.assertTrue(
