+++ temporian/implementation/numpy/operators/test/calendar_minute_test.py
@@ -25,7 +25,6 @@ from temporian.implementation.numpy.operators.calendar.minute import (
     CalendarMinuteNumpyImplementation,
 )
 from temporian.core.data import dtype
-from temporian.implementation.numpy.evaluator import run_with_check
 
 
 class CalendarMinuteNumpyImplementationTest(absltest.TestCase):
@@ -63,7 +62,7 @@ class CalendarMinuteNumpyImplementationTest(absltest.TestCase):
 
         operator = CalendarMinuteOperator(input_event)
         impl = CalendarMinuteNumpyImplementation(operator)
-        output = run_with_check(operator, impl, {"sampling": input_event_data})
+        output = impl.call(sampling=input_event_data)
 
         self.assertTrue(output_event_data == output["event"])
         self.assertTrue(
