+++ temporian/implementation/numpy/operators/test/calendar_second_test.py
@@ -25,7 +25,6 @@ from temporian.implementation.numpy.operators.calendar.second import (
     CalendarSecondNumpyImplementation,
 )
 from temporian.core.data import dtype
-from temporian.implementation.numpy.evaluator import run_with_check
 
 
 class CalendarSecondNumpyImplementationTest(absltest.TestCase):
@@ -62,7 +61,7 @@ class CalendarSecondNumpyImplementationTest(absltest.TestCase):
 
         operator = CalendarSecondOperator(input_event)
         impl = CalendarSecondNumpyImplementation(operator)
-        output = run_with_check(operator, impl, {"sampling": input_event_data})
+        output = impl.call(sampling=input_event_data)
 
         self.assertTrue(output_event_data == output["event"])
         self.assertTrue(
