+++ temporian/implementation/numpy/operators/test/lag_test.py
@@ -26,7 +26,7 @@ from temporian.core.operators.lag import leak
 from temporian.core.operators.lag import LagOperator
 from temporian.implementation.numpy.data.event import NumpyEvent
 from temporian.implementation.numpy.operators.lag import LagNumpyImplementation
-from temporian.implementation.numpy.evaluator import run_with_check
+
 from temporian.core.data import dtype as dtype_lib
 
 
@@ -79,9 +79,7 @@ class LagNumpyImplementationTest(absltest.TestCase):
         )
 
         lag_implementation = LagNumpyImplementation(operator)
-        operator_output = run_with_check(
-            operator, lag_implementation, {"event": numpy_input_event}
-        )
+        operator_output = lag_implementation.call(event=numpy_input_event)
 
         self.assertTrue(numpy_output_event == operator_output["event"])
 
@@ -223,9 +221,7 @@ class LagNumpyImplementationTest(absltest.TestCase):
         )
 
         lag_implementation = LagNumpyImplementation(operator)
-        operator_output = run_with_check(
-            operator, lag_implementation, {"event": numpy_input_event}
-        )
+        operator_output = lag_implementation.call(event=numpy_input_event)
 
         self.assertTrue(numpy_output_event == operator_output["event"])
