+++ temporian/implementation/numpy/operators/test/prefix_test.py
@@ -27,7 +27,6 @@ from temporian.implementation.numpy.data.sampling import NumpySampling
 from temporian.core.data import event as event_lib
 from temporian.core.data import feature as feature_lib
 from temporian.core.data import dtype as dtype_lib
-from temporian.implementation.numpy.evaluator import run_with_check
 
 
 class PrefixOperatorTest(absltest.TestCase):
@@ -60,7 +59,7 @@ class PrefixOperatorTest(absltest.TestCase):
         # Run op
         op = Prefix("hello_", event=event)
         instance = PrefixNumpyImplementation(op)
-        output = run_with_check(op, instance, {"event": event_data})["event"]
+        output = instance.call(event=event_data)["event"]
 
         self.assertEqual(output, expected_output)
